Fix naming of protobuf fields, add comments, fix formatting
