+++ src/contracts/index.ts
@@ -7,4 +7,4 @@ export { AddressMapper } from './address-mapper'
 export { TransferGateway } from './transfer-gateway'
 export { LoomCoinTransferGateway } from './loomcoin-gateway'
 export { TronTransferGateway } from './tron-gateway'
-export { UserDeployerWhitelist } from './user_deployer_whitelist'
+export { UserDeployerWhitelist } from './user-deployer-whitelist'
